Add custom security group 'allow_ssh' allowing ssh from all IPs
The default 'allow_ssh' security group was updated recently to limit
access to IP addresses originating on NVIDIA's internal network.
However, we need access to the instances we launch as part of the
kube-ci from the runners on public gitlab.

This patch is designed to address this issue. It adds a one-off security
group for the EC2 instance launch by the CI, that allows all IP
addresses to ssh into the instance.

Ideally we would limit it to a specific IP address (or range of IP
addresses), but gitlab doesn't provide such a list for its CI runners:

https://docs.gitlab.com/ee/user/gitlab_com/#ip-range

Signed-off-by: Kevin Klues <kklues@nvidia.com>