Support local creating a dev instance from terraform definition
This change allows the terraform definition to be used to create
a development instance. This is done by adding variables for the
key-pair name and associated private/public key pair and conditionally
creating the AWS resources based on these. The readme is also updated
to describe using this to create a development instance.

Signed-off-by: Evan Lezar <elezar@nvidia.com>